Simulating a pass through the prism and viewed expected downward deflection at each interface. Identified possible differences in deflection conditions/calculations due to the gradient of the prism sides. See inline comments in opticalFormulae.cpp

+++ opticalMedia.cpp
@@ -3,7 +3,7 @@
 #include "geometricUtils.h"
 
     
-Prism2D::Prism2D(int x, int y, int h, int n)
+Prism2D::Prism2D(int x, int y, int h, float n)
 {
     tipXPos = x;
     tipYPos = y;
